+++ example/main.dart
@@ -19,29 +19,6 @@ void main() async {
     ],
   );
 
-  final retryClient = Client(
-    middlewares: [
-      const LoggerMiddleware(logBody: true),
-      const HeadersMiddleware(headers: {'User-Agent': 'HttpToolkit/1.0'}),
-      RetryMiddleware(
-        maxRetries: 2,
-        strategy: const LinearBackoffStrategy(Duration(seconds: 1)),
-        whenError: (err, attempts, nextTry) {
-          print(
-            'Got Error in attempt $attempts, Retrying in ${nextTry.inMilliseconds}ms...',
-          );
-          return true;
-        },
-        whenResponse: (res, attempts, totalDuration) {
-          print(
-            'Got Response in attempt $attempts, Recovered in total ${totalDuration.inMilliseconds}ms...',
-          );
-          return true;
-        },
-      ),
-    ],
-  );
-
   try {
     final response = await client.get(
       Uri.parse('/todos/1'),
@@ -52,16 +29,7 @@ void main() async {
     }
   } on Exception catch (e) {
     print('Error: $e');
+  } finally {
+    client.close();
   }
-
-  try {
-    final response = await retryClient.get(
-      Uri.parse('http://localhost:8080'),
-    );
-    print('Status: ${response.statusCode}');
-  } on Exception catch (e) {
-    print('Error caught: $e');
-  }
-
-  client.close();
 }